refactor(useselector): replace connect(mapStateToProps)(...) with useSelector hook
note: clean use of state extraction via decomposition and removing work from useSelector.

Makes code much cleaner (compare to wprk in previous mapStateToProps, which to be faor was not related to
useSelector but to attempt to fine grade extraction of data.

This use of useSelector offers opportunity to write a very clean state extractor by
using decomposition, as in:

const { stateA,stateB } = useSelector ( state => state.someSliceOfState)